Cleanup & comments
